refactor: 핸들러에 섞여있던 뷰의 역할 분리
