+++ .eslintrc.json
@@ -10,15 +10,15 @@
   "ignorePatterns": ["cypress/"],
   "rules": {
     "no-new": "off",
-    "no-console": "off",
+    "no-console": "on",
     "no-alert": "off",
     "no-plusplus": "off",
     "consistent-return": "off",
     "no-param-reassign": "off",
     "no-return-assign": "off",
     "import/extensions": "off",
-    "import/prefer-default-export": "off"
-    // "max-depth": ["error", 1],
-    // "max-lines-per-function": ["error", 15]
+    "import/prefer-default-export": "off",
+    "max-depth": ["error", 1],
+    "max-lines-per-function": ["error", 15]
   }
 }
