+++ src/js/models/generateCarInstances.js
@@ -1,4 +1,4 @@
-import { Car } from '../class/Car.js';
+import { Car } from './Car.js';
 
 export const generateCarInstances = (carNameInputValue) => {
   return carNameInputValue.split(',').map((carName) => new Car(carName.trim()));
