+++ src/js/models/getWinners.js
