+++ src/js/views/alertGameOver.js
@@ -1,5 +1,5 @@
 import { GAME_OVER_NOTICE } from '../constants/gameOverNotice.js';
-import { wait } from './wait.js';
+import { wait } from '../utils/wait.js';
 
 export const alertGameOverAfterDelay = async (winners) => {
   const { DELAY, MSG_SUFFIX } = GAME_OVER_NOTICE;
