+++ src/js/views/showGameResult.js
@@ -1,6 +1,6 @@
 import { GAME_RESULT } from '../constants/gameOverNotice.js';
-import { $ } from './querySelector.js';
-import { setVisibility } from './setVisibility.js';
+import { $ } from '../utils/querySelector.js';
+import { setVisibility } from './utils/setVisibility.js';
 
 export const showGameResult = (winners) => {
   const { PREFIX, SUFFIX } = GAME_RESULT;
