+++ src/js/views/utils/setVisibility.js
